LayoutTests/imported/w3c/web-platform-tests/feature-policy/experimental-features/trust-token-redemption-default-feature-policy.tentative.https.sub-expected.txt

Blocked access to external URL https://www.localhost:9443/feature-policy/experimental-features/resources/feature-policy-trust-token-redemption.html


Harness Error (TIMEOUT), message = null

FAIL Default "trust-token-redemption" feature policy ["self"] allows the top-level document. assert_unreached: Reached unreachable code
FAIL Default "trust-token-redemption" feature policy ["self"] allows same-origin iframes. assert_equals: Trust token redemption expected 4 but got 2
TIMEOUT Default "trust-token-redemption" feature policy ["self"] disallows cross-origin iframes. Test timed out
